Moved test suites our of 'old' folder
