+++ core/test/acceptance/admin/actions_spec.js
@@ -1,9 +1,9 @@
 const should = require('should');
 const Promise = require('bluebird');
 const supertest = require('supertest');
-const testUtils = require('../../../utils');
+const testUtils = require('../../utils');
 const localUtils = require('./utils');
-const config = require('../../../../server/config');
+const config = require('../../../server/config');
 const ghost = testUtils.startGhost;
 
 let request;
