+++ core/test/acceptance/admin/config_spec.js
@@ -1,8 +1,8 @@
 const should = require('should');
 const supertest = require('supertest');
-const testUtils = require('../../../utils');
+const testUtils = require('../../utils');
 const localUtils = require('./utils');
-const config = require('../../../../server/config');
+const config = require('../../../server/config');
 const ghost = testUtils.startGhost;
 
 let request;
