+++ core/test/acceptance/admin/posts_spec.js
@@ -3,10 +3,11 @@ const supertest = require('supertest');
 const _ = require('lodash');
 const ObjectId = require('bson-objectid');
 const moment = require('moment-timezone');
-const testUtils = require('../../../utils');
+const testUtils = require('../../utils');
+const config = require('../../../server/config');
+const models = require('../../../server/models');
 const localUtils = require('./utils');
-const config = require('../../../../server/config');
-const models = require('../../../../server/models');
+
 const ghost = testUtils.startGhost;
 let request;
