+++ core/test/acceptance/admin/roles_spec.js
@@ -1,7 +1,7 @@
 const should = require('should');
 const supertest = require('supertest');
-const config = require('../../../../server/config');
-const testUtils = require('../../../utils');
+const config = require('../../../server/config');
+const testUtils = require('../../utils');
 const localUtils = require('./utils');
 
 const ghost = testUtils.startGhost;
