+++ core/test/acceptance/admin/settings_spec.js
@@ -3,8 +3,8 @@ const _ = require('lodash');
 const supertest = require('supertest');
 const os = require('os');
 const fs = require('fs-extra');
-const config = require('../../../../server/config');
-const testUtils = require('../../../utils');
+const config = require('../../../server/config');
+const testUtils = require('../../utils');
 const localUtils = require('./utils');
 const ghost = testUtils.startGhost;
 let request;
