+++ core/test/acceptance/admin/slugs_spec.js
@@ -1,8 +1,9 @@
 var should = require('should'),
     supertest = require('supertest'),
-    testUtils = require('../../../utils'),
-    localUtils = require('../../../acceptance/old/admin/utils'),
-    config = require('../../../../server/config'),
+    testUtils = require('../../utils'),
+    config = require('../../../server/config'),
+    localUtils = require('./utils'),
+
     ghost = testUtils.startGhost,
     request;
