+++ core/test/acceptance/content/authors_spec.js
@@ -2,11 +2,12 @@ const should = require('should');
 const supertest = require('supertest');
 const _ = require('lodash');
 const url = require('url');
-const configUtils = require('../../../utils/configUtils');
-const config = require('../../../../server/config');
-const models = require('../../../../server/models');
-const testUtils = require('../../../utils');
+const configUtils = require('../../utils/configUtils');
+const config = require('../../../server/config');
+const models = require('../../../server/models');
+const testUtils = require('../../utils');
 const localUtils = require('./utils');
+
 const ghost = testUtils.startGhost;
 let request;
