+++ core/test/acceptance/content/key_authentication_spec.js
@@ -1,9 +1,9 @@
 const should = require('should');
 const supertest = require('supertest');
 const _ = require('lodash');
-const testUtils = require('../../../utils');
+const testUtils = require('../../utils');
 const localUtils = require('./utils');
-const config = require('../../../../server/config');
+const config = require('../../../server/config');
 
 const ghost = testUtils.startGhost;
