+++ core/test/acceptance/content/pages_spec.js
@@ -1,10 +1,10 @@
 const url = require('url');
 const should = require('should');
 const supertest = require('supertest');
-const testUtils = require('../../../utils');
+const testUtils = require('../../utils');
+const configUtils = require('../../utils/configUtils');
+const config = require('../../../server/config');
 const localUtils = require('./utils');
-const configUtils = require('../../../utils/configUtils');
-const config = require('../../../../server/config');
 
 const ghost = testUtils.startGhost;
 let request;
