+++ core/test/acceptance/content/settings_spec.js
@@ -1,13 +1,13 @@
 const should = require('should');
 const supertest = require('supertest');
 const _ = require('lodash');
-const testUtils = require('../../../utils');
+const testUtils = require('../../utils');
+const config = require('../../../server/config');
 const localUtils = require('./utils');
-const config = require('../../../../server/config');
 
 // Values to test against
-const publicSettings = require('../../../../server/services/settings/public');
-const defaultSettings = require('../../../../server/data/schema').defaultSettings.blog;
+const publicSettings = require('../../../server/services/settings/public');
+const defaultSettings = require('../../../server/data/schema').defaultSettings.blog;
 
 const ghost = testUtils.startGhost;
 let request;
